fix(ai): remove MOCK_AI runtime flag — mock client tree-shaken out of prod bundle
Move MockAIService into lib/ai-service.mock.ts and acquire it only via a
NODE_ENV-guarded dynamic require, so a production Next build dead-code-
eliminates the mock module entirely. Update test/script imports to the new path.

Closes #19
