+++ recipe-lanes/tests/data.test.ts
@@ -2,7 +2,8 @@ import { describe, it, beforeEach } from 'node:test';
 import assert from 'node:assert';
 import { getDataService, setDataService, MemoryDataService } from '../lib/data-service';
 import { memoryStore } from '../lib/store';
-import { setAIService, MockAIService } from '../lib/ai-service';
+import { setAIService } from '../lib/ai-service';
+import { MockAIService } from '../lib/ai-service.mock';
 import { setAuthService, MockAuthService } from '../lib/auth-service';
 import { createVisualRecipeAction } from '../app/actions';
 import { getNodeIconUrl } from '../lib/recipe-lanes/model-utils';
