+++ recipe-lanes/tests/functions-metadata.test.ts
@@ -19,7 +19,8 @@ import { describe, it } from 'node:test';
 import assert from 'node:assert';
 import { db, storage } from '../lib/firebase-admin';
 import { createVisualRecipeAction, forgeIconAction } from '../app/actions';
-import { setAIService, MockAIService } from '../lib/ai-service';
+import { setAIService } from '../lib/ai-service';
+import { MockAIService } from '../lib/ai-service.mock';
 import { setAuthService, MockAuthService } from '../lib/auth-service';
 import { getIconPath } from '../lib/recipe-lanes/model-utils';
 import { standardizeIngredientName } from '../lib/utils';
