+++ recipe-lanes/tests/icon-index.test.ts
@@ -25,7 +25,8 @@
 
 import { describe, it, before } from 'node:test';
 import assert from 'node:assert';
-import { setAIService, MockAIService } from '../lib/ai-service';
+import { setAIService } from '../lib/ai-service';
+import { MockAIService } from '../lib/ai-service.mock';
 import { setAuthService, MockAuthService } from '../lib/auth-service';
 import { createDebugRecipeAction, addIngredientNodeAction } from '../app/actions';
 import { getDataService } from '../lib/data-service';
