+++ recipe-lanes/tests/icon-pipeline.test.ts
@@ -9,7 +9,8 @@ import { describe, it, beforeEach } from 'node:test';
 import assert from 'node:assert';
 import { getDataService, setDataService, MemoryDataService, DataService } from '../lib/data-service';
 import { memoryStore } from '../lib/store';
-import { setAIService, MockAIService } from '../lib/ai-service';
+import { setAIService } from '../lib/ai-service';
+import { MockAIService } from '../lib/ai-service.mock';
 import { setAuthService, MockAuthService } from '../lib/auth-service';
 import { buildShortlistEntry, getNodeIconUrl, getNodeIconId } from '../lib/recipe-lanes/model-utils';
 import type { IconStats, ShortlistEntry } from '../lib/recipe-lanes/types';
