+++ recipe-lanes/tests/impression-rejection.test.ts
@@ -30,7 +30,8 @@ import { describe, it } from 'node:test';
 import assert from 'node:assert/strict';
 import { db } from '../lib/firebase-admin';
 import { getDataService } from '../lib/data-service';
-import { setAIService, MockAIService } from '../lib/ai-service';
+import { setAIService } from '../lib/ai-service';
+import { MockAIService } from '../lib/ai-service.mock';
 import { setAuthService, MockAuthService } from '../lib/auth-service';
 import { buildShortlistEntry } from '../lib/recipe-lanes/model-utils';
 import { standardizeIngredientName } from '../lib/utils';
